Fix admin api build

+++ extra/admin-api/Spacebar.Db/Models/User.cs
@@ -96,7 +96,7 @@ public partial class User
     public string Flags { get; set; }
 
     [Column("public_flags")]
-    public int PublicFlags { get; set; }
+    public ulong PublicFlags { get; set; }
 
     [Column("purchased_flags")]
     public int PurchasedFlags { get; set; }
@@ -105,7 +105,7 @@ public partial class User
     public int PremiumUsageFlags { get; set; }
 
     [Column("rights")]
-    public long Rights { get; set; }
+    public ulong Rights { get; set; }
 
     [Column("data")]
     public string Data { get; set; } = null!;
